requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
Faker==20.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
